Move `entity_axe_knight` into shared header
